+++ tests/test_mosaic.py
@@ -15,7 +15,7 @@ from .conftest import mock_rasterio_open, parse_img
 def search_no_bbox(app):
     """Search Query without BBOX."""
     query = {"collections": ["noaa-emergency-response"], "filter-lang": "cql-json"}
-    response = app.post("/mosaic/register", json=query)
+    response = app.post("/mosaics/register", json=query)
     assert response.status_code == 200
     resp = response.json()
     assert resp["links"]
@@ -36,7 +36,7 @@ def search_bbox(app):
         "bbox": [-85.535, 36.137, -85.465, 36.179],
         "filter-lang": "cql-json",
     }
-    response = app.post("/mosaic/register", json=query)
+    response = app.post("/mosaics/register", json=query)
     assert response.status_code == 200
 
     resp = response.json()
@@ -52,7 +52,7 @@ def search_bbox(app):
 
 def test_info(app, search_no_bbox):
     """Should return metadata about a search query."""
-    response = app.get(f"/mosaic/{search_no_bbox}/info")
+    response = app.get(f"/mosaics/{search_no_bbox}/info")
     assert response.status_code == 200
     resp = response.json()
     assert resp["search"]
@@ -64,7 +64,7 @@ def test_info(app, search_no_bbox):
     }
     assert search["metadata"] == {"type": "mosaic"}
 
-    response = app.get("/mosaic/aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/info")
+    response = app.get("/mosaics/aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/info")
     assert response.status_code == 404
     resp = response.json()
     assert resp["detail"] == "SearchId `aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa` not found"
@@ -72,7 +72,7 @@ def test_info(app, search_no_bbox):
 
 def test_assets_for_point(app, search_no_bbox, search_bbox):
     """Get assets for a Point."""
-    response = app.get(f"/mosaic/{search_no_bbox}/-85.6358,36.1624/assets")
+    response = app.get(f"/mosaics/{search_no_bbox}/-85.6358,36.1624/assets")
     assert response.status_code == 200
     resp = response.json()
     assert len(resp) == 1
@@ -80,14 +80,14 @@ def test_assets_for_point(app, search_no_bbox, search_bbox):
     assert resp[0]["id"] == "20200307aC0853900w361030"
 
     # make sure we can find assets when having both bbox and geometry
-    response = app.get(f"/mosaic/{search_bbox}/-85.5,36.1624/assets")
+    response = app.get(f"/mosaics/{search_bbox}/-85.5,36.1624/assets")
     assert response.status_code == 200
     resp = response.json()
     assert len(resp) == 2
 
     # with coord-crs
     response = app.get(
-        f"/mosaic/{search_bbox}/-9517816.46282489,4322990.432036275/assets",
+        f"/mosaics/{search_bbox}/-9517816.46282489,4322990.432036275/assets",
         params={"coord_crs": "epsg:3857"},
     )
     assert response.status_code == 200
@@ -95,13 +95,13 @@ def test_assets_for_point(app, search_no_bbox, search_bbox):
     assert len(resp) == 2
 
     # no assets found outside the mosaic bbox
-    response = app.get(f"/mosaic/{search_bbox}/-85.6358,36.1624/assets")
+    response = app.get(f"/mosaics/{search_bbox}/-85.6358,36.1624/assets")
     assert response.status_code == 200
     resp = response.json()
     assert len(resp) == 0
 
     # searchId not found
-    response = app.get("/mosaic/aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/-85.5,36.1624/assets")
+    response = app.get("/mosaics/aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/-85.5,36.1624/assets")
     assert response.status_code == 404
     resp = response.json()
     assert resp["detail"] == "SearchId `aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa` not found"
@@ -109,7 +109,7 @@ def test_assets_for_point(app, search_no_bbox, search_bbox):
 
 def test_assets_for_tile(app, search_no_bbox, search_bbox):
     """Get assets for a Tile."""
-    response = app.get(f"/mosaic/{search_no_bbox}/tiles/15/8589/12849/assets")
+    response = app.get(f"/mosaics/{search_no_bbox}/tiles/15/8589/12849/assets")
     assert response.status_code == 200
     resp = response.json()
     assert len(resp) == 1
@@ -117,38 +117,38 @@ def test_assets_for_tile(app, search_no_bbox, search_bbox):
     assert resp[0]["id"] == "20200307aC0853900w361030"
 
     # make sure we can find assets when having both bbox and geometry
-    response = app.get(f"/mosaic/{search_bbox}/tiles/15/8601/12849/assets")
+    response = app.get(f"/mosaics/{search_bbox}/tiles/15/8601/12849/assets")
     assert response.status_code == 200
     resp = response.json()
     assert len(resp) == 2
 
     response = app.get(
-        f"/mosaic/{search_bbox}/tiles/WebMercatorQuad/15/8601/12849/assets"
+        f"/mosaics/{search_bbox}/tiles/WebMercatorQuad/15/8601/12849/assets"
     )
     assert response.status_code == 200
     resp = response.json()
     assert len(resp) == 2
 
     # With WGS1994Quad TMS
-    response = app.get(f"/mosaic/{search_bbox}/tiles/WGS1984Quad/14/8601/4901/assets")
+    response = app.get(f"/mosaics/{search_bbox}/tiles/WGS1984Quad/14/8601/4901/assets")
     assert response.status_code == 200
     resp = response.json()
     assert len(resp) == 4
 
-    response = app.get(f"/mosaic/{search_bbox}/tiles/15/8601/12849/assets")
+    response = app.get(f"/mosaics/{search_bbox}/tiles/15/8601/12849/assets")
     assert response.status_code == 200
     resp = response.json()
     assert len(resp) == 2
 
     # no assets found outside the query bbox
-    response = app.get(f"/mosaic/{search_bbox}/tiles/15/8589/12849/assets")
+    response = app.get(f"/mosaics/{search_bbox}/tiles/15/8589/12849/assets")
     assert response.status_code == 200
     resp = response.json()
     assert len(resp) == 0
 
     # searchId not found
     response = app.get(
-        "/mosaic/aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/tiles/15/8589/12849/assets"
+        "/mosaics/aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/tiles/15/8589/12849/assets"
     )
     assert response.status_code == 404
     resp = response.json()
@@ -157,10 +157,10 @@ def test_assets_for_tile(app, search_no_bbox, search_bbox):
 
 def test_tilejson(app, search_no_bbox, search_bbox):
     """Create TileJSON."""
-    response = app.get(f"/mosaic/{search_no_bbox}/tilejson.json")
+    response = app.get(f"/mosaics/{search_no_bbox}/tilejson.json")
     assert response.status_code == 400
 
-    response = app.get(f"/mosaic/{search_no_bbox}/tilejson.json?assets=cog")
+    response = app.get(f"/mosaics/{search_no_bbox}/tilejson.json?assets=cog")
     assert response.headers["content-type"] == "application/json"
     assert response.status_code == 200
     resp = response.json()
@@ -171,7 +171,7 @@ def test_tilejson(app, search_no_bbox, search_bbox):
     assert "?assets=cog" in resp["tiles"][0]
 
     response = app.get(
-        f"/mosaic/{search_no_bbox}/tilejson.json?assets=cog&scan_limit=100&items_limit=1&time_limit=2&exitwhenfull=False&skipcovered=False"
+        f"/mosaics/{search_no_bbox}/tilejson.json?assets=cog&scan_limit=100&items_limit=1&time_limit=2&exitwhenfull=False&skipcovered=False"
     )
     assert response.headers["content-type"] == "application/json"
     assert response.status_code == 200
@@ -181,18 +181,18 @@ def test_tilejson(app, search_no_bbox, search_bbox):
         in resp["tiles"][0]
     )
 
-    response = app.get(f"/mosaic/{search_no_bbox}/tilejson.json?expression=cog")
+    response = app.get(f"/mosaics/{search_no_bbox}/tilejson.json?expression=cog")
     assert response.status_code == 200
     resp = response.json()
     assert "?expression=cog" in resp["tiles"][0]
 
-    response = app.get(f"/mosaic/{search_no_bbox}/tilejson.json?expression=cog")
+    response = app.get(f"/mosaics/{search_no_bbox}/tilejson.json?expression=cog")
     assert response.status_code == 200
     resp = response.json()
     assert "?expression=cog" in resp["tiles"][0]
 
     response = app.get(
-        f"/mosaic/{search_no_bbox}/WorldCRS84Quad/tilejson.json?assets=cog"
+        f"/mosaics/{search_no_bbox}/WorldCRS84Quad/tilejson.json?assets=cog"
     )
     assert response.status_code == 200
     resp = response.json()
@@ -203,13 +203,13 @@ def test_tilejson(app, search_no_bbox, search_bbox):
     assert "?assets=cog" in resp["tiles"][0]
 
     response = app.get(
-        f"/mosaic/{search_no_bbox}/tilejson.json?assets=cog&tile_format=png"
+        f"/mosaics/{search_no_bbox}/tilejson.json?assets=cog&tile_format=png"
     )
     assert response.status_code == 200
     resp = response.json()
     assert ".png?assets=cog" in resp["tiles"][0]
 
-    response = app.get(f"/mosaic/{search_bbox}/tilejson.json?assets=cog")
+    response = app.get(f"/mosaics/{search_bbox}/tilejson.json?assets=cog")
     assert response.headers["content-type"] == "application/json"
     assert response.status_code == 200
     resp = response.json()
@@ -220,7 +220,7 @@ def test_tilejson(app, search_no_bbox, search_bbox):
     assert "?assets=cog" in resp["tiles"][0]
 
     # searchId not found
-    response = app.get("/mosaic/aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/info?assets=cog")
+    response = app.get("/mosaics/aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/info?assets=cog")
     assert response.status_code == 404
     resp = response.json()
     assert resp["detail"] == "SearchId `aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa` not found"
@@ -234,10 +234,10 @@ def test_tiles(rio, app, search_no_bbox, search_bbox):
     z, x, y = 15, 8589, 12849
 
     # missing assets
-    response = app.get(f"/mosaic/{search_no_bbox}/tiles/{z}/{x}/{y}")
+    response = app.get(f"/mosaics/{search_no_bbox}/tiles/{z}/{x}/{y}")
     assert response.status_code == 400
 
-    response = app.get(f"/mosaic/{search_no_bbox}/tiles/{z}/{x}/{y}?assets=cog")
+    response = app.get(f"/mosaics/{search_no_bbox}/tiles/{z}/{x}/{y}?assets=cog")
     assert response.status_code == 200
     assert response.headers["content-type"] == "image/jpeg"
     meta = parse_img(response.content)
@@ -245,7 +245,7 @@ def test_tiles(rio, app, search_no_bbox, search_bbox):
     assert meta["height"] == 256
 
     response = app.get(
-        f"/mosaic/{search_no_bbox}/tiles/{z}/{x}/{y}?assets=cog&buffer=0.5"
+        f"/mosaics/{search_no_bbox}/tiles/{z}/{x}/{y}?assets=cog&buffer=0.5"
     )
     assert response.status_code == 200
     assert response.headers["content-type"] == "image/jpeg"
@@ -253,7 +253,7 @@ def test_tiles(rio, app, search_no_bbox, search_bbox):
     assert meta["width"] == 257
     assert meta["height"] == 257
 
-    response = app.get(f"/mosaic/{search_no_bbox}/tiles/{z}/{x}/{y}.png?assets=cog")
+    response = app.get(f"/mosaics/{search_no_bbox}/tiles/{z}/{x}/{y}.png?assets=cog")
     assert response.status_code == 200
     assert response.headers["content-type"] == "image/png"
     meta = parse_img(response.content)
@@ -261,11 +261,11 @@ def test_tiles(rio, app, search_no_bbox, search_bbox):
     assert meta["height"] == 256
 
     # tile is outside mosaic bbox, it should return 404 (NoAssetFoundError)
-    response = app.get(f"/mosaic/{search_bbox}/tiles/{z}/{x}/{y}?assets=cog")
+    response = app.get(f"/mosaics/{search_bbox}/tiles/{z}/{x}/{y}?assets=cog")
     assert response.status_code == 404
 
     response = app.get(
-        f"/mosaic/{search_no_bbox}/tiles/WebMercatorQuad/{z}/{x}/{y}.tif?assets=cog"
+        f"/mosaics/{search_no_bbox}/tiles/WebMercatorQuad/{z}/{x}/{y}.tif?assets=cog"
     )
     assert response.status_code == 200
     assert response.headers["content-type"] == "image/tiff; application=geotiff"
@@ -275,7 +275,7 @@ def test_tiles(rio, app, search_no_bbox, search_bbox):
     assert meta["height"] == 256
 
     response = app.get(
-        f"/mosaic/{search_no_bbox}/tiles/WorldCRS84Quad/18/137421/78424.tif?assets=cog"
+        f"/mosaics/{search_no_bbox}/tiles/WorldCRS84Quad/18/137421/78424.tif?assets=cog"
     )
     assert response.status_code == 200
     assert response.headers["content-type"] == "image/tiff; application=geotiff"
@@ -286,7 +286,7 @@ def test_tiles(rio, app, search_no_bbox, search_bbox):
 
     # searchId not found
     response = app.get(
-        "/mosaic/aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/tiles/0/0/0?assets=cog"
+        "/mosaics/aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/tiles/0/0/0?assets=cog"
     )
     assert response.status_code == 404
     resp = response.json()
@@ -296,14 +296,14 @@ def test_tiles(rio, app, search_no_bbox, search_bbox):
 def test_wmts(app, search_no_bbox):
     """Create wmts document."""
     # missing assets
-    response = app.get(f"/mosaic/{search_no_bbox}/WMTSCapabilities.xml")
+    response = app.get(f"/mosaics/{search_no_bbox}/WMTSCapabilities.xml")
     assert response.status_code == 400
     assert (
         response.json()["detail"]
         == "assets must be defined either via expression or assets options."
     )
 
-    response = app.get(f"/mosaic/{search_no_bbox}/WMTSCapabilities.xml?assets=cog")
+    response = app.get(f"/mosaics/{search_no_bbox}/WMTSCapabilities.xml?assets=cog")
     assert response.status_code == 200
     assert response.headers["content-type"] == "application/xml"
 
@@ -313,7 +313,7 @@ def test_wmts(app, search_no_bbox):
         assert src.profile["driver"] == "WMTS"
 
     response = app.get(
-        f"/mosaic/{search_no_bbox}/WorldCRS84Quad/WMTSCapabilities.xml?assets=cog"
+        f"/mosaics/{search_no_bbox}/WorldCRS84Quad/WMTSCapabilities.xml?assets=cog"
     )
     assert response.status_code == 200
     assert response.headers["content-type"] == "application/xml"
@@ -335,7 +335,7 @@ def test_cql2(rio, app):
             "args": [{"property": "collection"}, "noaa-emergency-response"],
         }
     }
-    response = app.post("/mosaic/register", json=query)
+    response = app.post("/mosaics/register", json=query)
     assert response.status_code == 200
     resp = response.json()
     assert resp["id"]
@@ -343,7 +343,7 @@ def test_cql2(rio, app):
 
     cql2_id = resp["id"]
 
-    response = app.get(f"/mosaic/{cql2_id}/info")
+    response = app.get(f"/mosaics/{cql2_id}/info")
     assert response.status_code == 200
     resp = response.json()
     assert resp["search"]
@@ -357,21 +357,21 @@ def test_cql2(rio, app):
     }
     assert search["metadata"] == {"type": "mosaic"}
 
-    response = app.get(f"/mosaic/{cql2_id}/-85.6358,36.1624/assets")
+    response = app.get(f"/mosaics/{cql2_id}/-85.6358,36.1624/assets")
     assert response.status_code == 200
     resp = response.json()
     assert len(resp) == 1
     assert list(resp[0]) == ["id", "bbox", "assets", "collection"]
     assert resp[0]["id"] == "20200307aC0853900w361030"
 
-    response = app.get(f"/mosaic/{cql2_id}/tiles/15/8589/12849/assets")
+    response = app.get(f"/mosaics/{cql2_id}/tiles/15/8589/12849/assets")
     assert response.status_code == 200
     resp = response.json()
     assert len(resp) == 1
     assert list(resp[0]) == ["id", "bbox", "assets", "collection"]
     assert resp[0]["id"] == "20200307aC0853900w361030"
 
-    response = app.get(f"/mosaic/{cql2_id}/tilejson.json?assets=cog")
+    response = app.get(f"/mosaics/{cql2_id}/tilejson.json?assets=cog")
     assert response.headers["content-type"] == "application/json"
     assert response.status_code == 200
     resp = response.json()
@@ -381,12 +381,12 @@ def test_cql2(rio, app):
     assert round(resp["bounds"][0]) == -180
     # Make sure we return a tilejson with the `/{search_id}/tiles/{tms}` format
     assert (
-        f"/mosaic/{cql2_id}/tiles/WebMercatorQuad/{{z}}/{{x}}/{{y}}?assets=cog"
+        f"/mosaics/{cql2_id}/tiles/WebMercatorQuad/{{z}}/{{x}}/{{y}}?assets=cog"
         in resp["tiles"][0]
     )
 
     z, x, y = 15, 8589, 12849
-    response = app.get(f"/mosaic/{cql2_id}/tiles/{z}/{x}/{y}?assets=cog")
+    response = app.get(f"/mosaics/{cql2_id}/tiles/{z}/{x}/{y}?assets=cog")
     assert response.status_code == 200
     assert response.headers["content-type"] == "image/jpeg"
     meta = parse_img(response.content)
@@ -428,7 +428,7 @@ def test_cql2_with_geometry(rio, app):
             ],
         }
     }
-    response = app.post("/mosaic/register", json=query)
+    response = app.post("/mosaics/register", json=query)
     assert response.status_code == 200
     resp = response.json()
     assert resp["id"]
@@ -436,7 +436,7 @@ def test_cql2_with_geometry(rio, app):
 
     cql2_id = resp["id"]
 
-    response = app.get(f"/mosaic/{cql2_id}/info")
+    response = app.get(f"/mosaics/{cql2_id}/info")
     assert response.status_code == 200
     resp = response.json()
     assert resp["search"]
@@ -445,32 +445,32 @@ def test_cql2_with_geometry(rio, app):
     assert search["metadata"] == {"type": "mosaic"}
 
     # make sure we can find assets when having both geometry filter and geometry
-    response = app.get(f"/mosaic/{cql2_id}/tiles/15/8601/12849/assets")
+    response = app.get(f"/mosaics/{cql2_id}/tiles/15/8601/12849/assets")
     assert response.status_code == 200
     resp = response.json()
     assert len(resp) == 2
 
     # point is outside the geometry filter
-    response = app.get(f"/mosaic/{cql2_id}/-85.6358,36.1624/assets")
+    response = app.get(f"/mosaics/{cql2_id}/-85.6358,36.1624/assets")
     assert response.status_code == 200
     resp = response.json()
     assert len(resp) == 0
 
     # make sure we can find assets when having both geometry filter and geometry
-    response = app.get(f"/mosaic/{cql2_id}/tiles/15/8601/12849/assets")
+    response = app.get(f"/mosaics/{cql2_id}/tiles/15/8601/12849/assets")
     assert response.status_code == 200
     resp = response.json()
     assert len(resp) == 2
 
     # tile is outside the geometry filter
-    response = app.get(f"/mosaic/{cql2_id}/tiles/15/8589/12849/assets")
+    response = app.get(f"/mosaics/{cql2_id}/tiles/15/8589/12849/assets")
     assert response.status_code == 200
     resp = response.json()
     assert len(resp) == 0
 
     # tile is outside the geometry filter
     z, x, y = 15, 8589, 12849
-    response = app.get(f"/mosaic/{cql2_id}/tiles/{z}/{x}/{y}?assets=cog")
+    response = app.get(f"/mosaics/{cql2_id}/tiles/{z}/{x}/{y}?assets=cog")
     assert response.status_code == 404
 
 
@@ -484,7 +484,7 @@ def test_query_with_metadata(app):
         "metadata": {"name": "mymosaic", "minzoom": 1, "maxzoom": 2},
     }
 
-    response = app.post("/mosaic/register", json=query)
+    response = app.post("/mosaics/register", json=query)
     assert response.status_code == 200
     resp = response.json()
     assert resp["id"]
@@ -492,7 +492,7 @@ def test_query_with_metadata(app):
 
     mosaic_id = resp["id"]
 
-    response = app.get(f"/mosaic/{mosaic_id}/info")
+    response = app.get(f"/mosaics/{mosaic_id}/info")
     assert response.status_code == 200
     resp = response.json()
     assert resp["search"]
@@ -511,7 +511,7 @@ def test_query_with_metadata(app):
         "maxzoom": 2,
     }
 
-    response = app.get(f"/mosaic/{mosaic_id}/tilejson.json?assets=cog")
+    response = app.get(f"/mosaics/{mosaic_id}/tilejson.json?assets=cog")
     assert response.status_code == 200
     resp = response.json()
     assert resp["minzoom"] == 1
@@ -545,7 +545,7 @@ def test_query_with_metadata(app):
     }
 
     with pytest.warns(UserWarning):
-        response = app.post("/mosaic/register", json=query)
+        response = app.post("/mosaics/register", json=query)
     assert response.status_code == 200
     resp = response.json()
     assert resp["id"]
@@ -562,7 +562,7 @@ def test_query_with_metadata(app):
 
     # Test WMTS
     # 1. missing assets and no metadata layers
-    response = app.get(f"/mosaic/{mosaic_id}/WMTSCapabilities.xml")
+    response = app.get(f"/mosaics/{mosaic_id}/WMTSCapabilities.xml")
     assert response.status_code == 400
     assert (
         response.json()["detail"]
@@ -571,7 +571,7 @@ def test_query_with_metadata(app):
 
     # 2. assets and no metadata layers
     response = app.get(
-        f"/mosaic/{mosaic_id}/WMTSCapabilities.xml", params={"assets": "cog"}
+        f"/mosaics/{mosaic_id}/WMTSCapabilities.xml", params={"assets": "cog"}
     )
     assert response.status_code == 200
     assert response.headers["content-type"] == "application/xml"
@@ -583,7 +583,7 @@ def test_query_with_metadata(app):
 
     # 3. no assets and metadata layers
     with pytest.warns(UserWarning):
-        response = app.get(f"/mosaic/{mosaic_id_metadata}/WMTSCapabilities.xml")
+        response = app.get(f"/mosaics/{mosaic_id_metadata}/WMTSCapabilities.xml")
     assert response.status_code == 200
 
     assert response.headers["content-type"] == "application/xml"
@@ -597,7 +597,7 @@ def test_query_with_metadata(app):
     # 4. assets and metadata layers
     with pytest.warns(UserWarning):
         response = app.get(
-            f"/mosaic/{mosaic_id_metadata}/WMTSCapabilities.xml",
+            f"/mosaics/{mosaic_id_metadata}/WMTSCapabilities.xml",
             params={"assets": "cog"},
         )
     assert response.status_code == 200
@@ -609,6 +609,20 @@ def test_query_with_metadata(app):
         assert src.subdatasets[0].endswith(",layer=one_band")
         assert src.subdatasets[1].endswith(",layer=three_bands")
         assert src.subdatasets[2].endswith(",layer=default")
+
+    with pytest.warns(UserWarning):
+        response = app.get(f"/mosaics/{mosaic_id_metadata}/info")
+    assert response.status_code == 200
+    resp = response.json()
+    assert resp["search"]["hash"] == mosaic_id_metadata
+    assert len(resp["links"]) == 10  # self, tilejson (3), map (3), wmts (3)
+
+    assert resp["links"][1]["title"] == "TileJSON link (Template URL)"
+    assert resp["links"][2]["title"] == "TileJSON link for `one_band` layer"
+    assert resp["links"][3]["title"] == "TileJSON link for `three_bands` layer"
+
+    assert "asset_bidx=cog%7C1" in resp["links"][2]["href"]
+    assert "assets=cog" in resp["links"][2]["href"]
 
 
 @patch("rio_tiler.io.rasterio.rasterio")
@@ -643,12 +657,12 @@ def test_statistics(rio, app, search_no_bbox, search_bbox):
     }
 
     response = app.post(
-        f"/mosaic/{search_no_bbox}/statistics", json=feat, params={"max_size": 1024}
+        f"/mosaics/{search_no_bbox}/statistics", json=feat, params={"max_size": 1024}
     )
     assert response.status_code == 400
 
     response = app.post(
-        f"/mosaic/{search_no_bbox}/statistics",
+        f"/mosaics/{search_no_bbox}/statistics",
         json=feat,
         params={"assets": "cog", "max_size": 1024},
     )
@@ -657,7 +671,7 @@ def test_statistics(rio, app, search_no_bbox, search_bbox):
     assert response.json()["features"][0]["properties"]["statistics"]["cog_b1"]
 
     response = app.post(
-        f"/mosaic/{search_no_bbox}/statistics",
+        f"/mosaics/{search_no_bbox}/statistics",
         json=feat["features"][0],
         params={"assets": "cog", "max_size": 1024},
     )
@@ -667,7 +681,7 @@ def test_statistics(rio, app, search_no_bbox, search_bbox):
 
     # searchId not found
     response = app.post(
-        "/mosaic/aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/statistics",
+        "/mosaics/aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/statistics",
         json=feat,
         params={"assets": "cog", "max_size": 1024},
     )
@@ -678,21 +692,21 @@ def test_statistics(rio, app, search_no_bbox, search_bbox):
 
 def test_mosaic_list(app):
     """Test list mosaic."""
-    response = app.get("/mosaic/list")
+    response = app.get("/mosaics/list")
     assert response.status_code == 200
     resp = response.json()
     assert ["searches", "links", "context"] == list(resp)
     assert len(resp["searches"]) > 0
     assert len(resp["links"]) == 1
 
-    response = app.get("/mosaic/list?limit=1")
+    response = app.get("/mosaics/list?limit=1")
     assert response.status_code == 200
     resp = response.json()
     assert ["searches", "links", "context"] == list(resp)
     assert len(resp["searches"]) == 1
     assert len(resp["links"]) == 2
 
-    response = app.get("/mosaic/list?limit=1&offset=1")
+    response = app.get("/mosaics/list?limit=1&offset=1")
     assert response.status_code == 200
     resp = response.json()
     assert ["searches", "links", "context"] == list(resp)
@@ -706,7 +720,7 @@ def test_mosaic_list(app):
         },
         "metadata": {"data": "noaa", "num": 2},
     }
-    response = app.post("/mosaic/register", json=query)
+    response = app.post("/mosaics/register", json=query)
 
     query = {
         "filter": {
@@ -715,7 +729,7 @@ def test_mosaic_list(app):
         },
         "metadata": {"data": "noaa", "num": 1},
     }
-    response = app.post("/mosaic/register", json=query)
+    response = app.post("/mosaics/register", json=query)
 
     query = {
         "filter": {
@@ -724,9 +738,9 @@ def test_mosaic_list(app):
         },
         "metadata": {"data": "noaa", "num": 3},
     }
-    response = app.post("/mosaic/register", json=query)
+    response = app.post("/mosaics/register", json=query)
 
-    response = app.get("/mosaic/list?data=noaa")
+    response = app.get("/mosaics/list?data=noaa")
     assert response.status_code == 200
     resp = response.json()
     assert ["searches", "links", "context"] == list(resp)
@@ -734,22 +748,22 @@ def test_mosaic_list(app):
     assert len(resp["searches"]) == 3
     assert [s["search"]["metadata"]["num"] for s in resp["searches"]] == [2, 1, 3]
 
-    response = app.get("/mosaic/list?data=noaa&sortby=%2Bnum")
+    response = app.get("/mosaics/list?data=noaa&sortby=%2Bnum")
     assert response.status_code == 200
     resp = response.json()
     assert [s["search"]["metadata"]["num"] for s in resp["searches"]] == [1, 2, 3]
 
-    response = app.get("/mosaic/list?data=noaa&sortby=num")
+    response = app.get("/mosaics/list?data=noaa&sortby=num")
     assert response.status_code == 200
     resp = response.json()
     assert [s["search"]["metadata"]["num"] for s in resp["searches"]] == [1, 2, 3]
 
-    response = app.get("/mosaic/list?data=noaa&sortby=-num")
+    response = app.get("/mosaics/list?data=noaa&sortby=-num")
     assert response.status_code == 200
     resp = response.json()
     assert [s["search"]["metadata"]["num"] for s in resp["searches"]] == [3, 2, 1]
 
-    response = app.get("/mosaic/list?sortby=lastused")
+    response = app.get("/mosaics/list?sortby=lastused")
     assert response.status_code == 200
     resp = response.json()
     assert resp["context"]
@@ -759,7 +773,7 @@ def test_mosaic_list(app):
     ]
     assert dates[0] < dates[-1]
 
-    response = app.get("/mosaic/list?sortby=-lastused")
+    response = app.get("/mosaics/list?sortby=-lastused")
     assert response.status_code == 200
     resp = response.json()
     assert resp["context"]
@@ -772,10 +786,10 @@ def test_mosaic_list(app):
 
 def test_map(app, search_bbox):
     """test /map endpoint."""
-    response = app.get(f"/mosaic/{search_bbox}/map")
+    response = app.get(f"/mosaics/{search_bbox}/map")
     assert response.status_code == 400
 
-    response = app.get(f"/mosaic/{search_bbox}/map", params={"assets": "cog"})
+    response = app.get(f"/mosaics/{search_bbox}/map", params={"assets": "cog"})
     assert response.status_code == 200
 
 
@@ -806,12 +820,12 @@ def test_feature(rio, app, search_no_bbox):
     }
 
     response = app.post(
-        f"/mosaic/{search_no_bbox}/feature", json=feat, params={"max_size": 1024}
+        f"/mosaics/{search_no_bbox}/feature", json=feat, params={"max_size": 1024}
     )
     assert response.status_code == 400
 
     response = app.post(
-        f"/mosaic/{search_no_bbox}/feature",
+        f"/mosaics/{search_no_bbox}/feature",
         json=feat,
         params={"assets": "cog", "max_size": 1024},
     )
@@ -823,7 +837,7 @@ def test_feature(rio, app, search_no_bbox):
     assert meta["count"] == 4
 
     response = app.post(
-        f"/mosaic/{search_no_bbox}/feature.jpeg",
+        f"/mosaics/{search_no_bbox}/feature.jpeg",
         json=feat,
         params={"assets": "cog", "max_size": 1024},
     )
@@ -835,7 +849,7 @@ def test_feature(rio, app, search_no_bbox):
     assert meta["count"] == 3
 
     response = app.post(
-        f"/mosaic/{search_no_bbox}/feature/300x400.jpeg",
+        f"/mosaics/{search_no_bbox}/feature/300x400.jpeg",
         json=feat,
         params={"assets": "cog", "max_size": 1024},
     )
@@ -847,7 +861,7 @@ def test_feature(rio, app, search_no_bbox):
     assert meta["count"] == 3
 
     response = app.post(
-        f"/mosaic/{search_no_bbox}/feature/300x400.tif",
+        f"/mosaics/{search_no_bbox}/feature/300x400.tif",
         json=feat,
         params={"assets": "cog", "max_size": 1024},
     )
@@ -857,7 +871,7 @@ def test_feature(rio, app, search_no_bbox):
     assert meta["crs"] == "epsg:4326"
 
     response = app.post(
-        f"/mosaic/{search_no_bbox}/feature/300x400.tif",
+        f"/mosaics/{search_no_bbox}/feature/300x400.tif",
         json=feat,
         params={"assets": "cog", "max_size": 1024, "dst_crs": "epsg:3857"},
     )
@@ -880,12 +894,12 @@ def test_bbox(rio, app, search_no_bbox):
     ]
     str_bbox = ",".join(map(str, bbox))
     response = app.get(
-        f"/mosaic/{search_no_bbox}/bbox/{str_bbox}.png", params={"max_size": 1024}
+        f"/mosaics/{search_no_bbox}/bbox/{str_bbox}.png", params={"max_size": 1024}
     )
     assert response.status_code == 400
 
     response = app.get(
-        f"/mosaic/{search_no_bbox}/bbox/{str_bbox}.png",
+        f"/mosaics/{search_no_bbox}/bbox/{str_bbox}.png",
         params={"assets": "cog", "max_size": 1024},
     )
     assert response.status_code == 200
@@ -896,7 +910,7 @@ def test_bbox(rio, app, search_no_bbox):
     assert meta["count"] == 4
 
     response = app.get(
-        f"/mosaic/{search_no_bbox}/bbox/{str_bbox}.jpeg",
+        f"/mosaics/{search_no_bbox}/bbox/{str_bbox}.jpeg",
         params={"assets": "cog", "max_size": 1024},
     )
     assert response.status_code == 200
@@ -907,7 +921,7 @@ def test_bbox(rio, app, search_no_bbox):
     assert meta["count"] == 3
 
     response = app.get(
-        f"/mosaic/{search_no_bbox}/bbox/{str_bbox}/300x400.jpeg",
+        f"/mosaics/{search_no_bbox}/bbox/{str_bbox}/300x400.jpeg",
         params={"assets": "cog", "max_size": 1024},
     )
     assert response.status_code == 200
@@ -918,7 +932,7 @@ def test_bbox(rio, app, search_no_bbox):
     assert meta["count"] == 3
 
     response = app.get(
-        f"/mosaic/{search_no_bbox}/bbox/{str_bbox}.tif",
+        f"/mosaics/{search_no_bbox}/bbox/{str_bbox}.tif",
         params={"assets": "cog", "max_size": 1024},
     )
     assert response.status_code == 200
@@ -927,7 +941,7 @@ def test_bbox(rio, app, search_no_bbox):
     assert meta["crs"] == "epsg:4326"
 
     response = app.get(
-        f"/mosaic/{search_no_bbox}/bbox/{str_bbox}.tif",
+        f"/mosaics/{search_no_bbox}/bbox/{str_bbox}.tif",
         params={"assets": "cog", "max_size": 1024, "dst_crs": "epsg:3857"},
     )
     assert response.status_code == 200