+++ CHANGES.md
@@ -40,6 +40,8 @@
     app.include_router(mosaic.router, tags=["Mosaic"])
     ```
 
+* add *metadata layers* links in `/info` response for TileJSON, map and wmts endpoint links.
+
 ## 0.8.0 (2023-10-06)
 
 * update titiler requirement to `>=0.14.0,<0.15`